M5/requirements.txt
attrs==23.1.0
bcrypt==4.0.1
beautifulsoup4==4.12.2
blinker==1.7.0
certifi==2023.11.17
charset-normalizer==3.3.2
click==8.1.3
colorama==0.4.6
dnspython==2.4.2
email-validator==2.1.0.post1
Flask==2.2.3
Flask-Bcrypt==1.0.1
Flask-Login==0.6.3
Flask-Principal==0.4.0
Flask-WTF==1.2.1
gunicorn==20.1.0
idna==3.4
importlib-metadata==5.0.0
iniconfig==2.0.0
itsdangerous==2.1.2
Jinja2==3.1.2
jsons==1.6.3
MarkupSafe==2.1.1
mysql-connector-python==8.2.0
numpy==1.26.1
packaging==23.2
pluggy==1.3.0
protobuf==4.21.12
py==1.11.0
pytest==7.1.3
python-dotenv==1.0.0
requests==2.31.0
setuptools==68.2.2
soupsieve==2.5
tomli==2.0.1
typish==1.9.3
urllib3==2.1.0
Werkzeug==2.2.3
WTForms==3.1.1
zipp==3.9.0
